pyproject.toml
[tool.poetry]
name = "language app user management"
version = "0.1.0"
description = ""
authors = ["Arthur Khe <arthur4ik.khe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.1"}
motor = "^3.1.2"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
mongomock = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
